backend/requirements.txt
Django>=4.2,<4.3
djangorestframework>=3.15,<3.17
django-cors-headers>=4.5,<4.9
